# swagger.yml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Test
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /result/edit/comment/{skillIndex}/{commentIndex}:
    put:
      tags:
        - result-controller
      summary: изменить комментарий
      description: меняет комментарий под нужным индексом
      operationId: editComment
      parameters:
        - name: skillIndex
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: commentIndex
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentEditRequestDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
...
  /result:
    post:
      tags:
        - result-controller
      summary: Добавить результат
      description: Добавляет результат для прохождения теста.
      operationId: addResult
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResultRequestDTO'
        required: true
      responses:
        '201':
          description: Результат успешно добавлен
          content:
            '*/*':
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
        '400':
          description: Некорректные данные запроса
          content:
            '*/*':
              schema:
                type: string
                enum:
                  - 100 CONTINUE
                  - 101 SWITCHING_PROTOCOLS
                  - 102 PROCESSING
                  - 103 EARLY_HINTS
                  - 103 CHECKPOINT
                  - 200 OK
                  - 201 CREATED
                  - 202 ACCEPTED
                  - 203 NON_AUTHORITATIVE_INFORMATION
                  - 204 NO_CONTENT
                  - 205 RESET_CONTENT
                  - 206 PARTIAL_CONTENT
                  - 207 MULTI_STATUS
                  - 208 ALREADY_REPORTED
                  - 226 IM_USED
                  - 300 MULTIPLE_CHOICES
                  - 301 MOVED_PERMANENTLY
                  - 302 FOUND
                  - 302 MOVED_TEMPORARILY
                  - 303 SEE_OTHER
                  - 304 NOT_MODIFIED
                  - 305 USE_PROXY
                  - 307 TEMPORARY_REDIRECT
                  - 308 PERMANENT_REDIRECT
                  - 400 BAD_REQUEST
                  - 401 UNAUTHORIZED
                  - 402 PAYMENT_REQUIRED
                  - 403 FORBIDDEN
                  - 404 NOT_FOUND
                  - 405 METHOD_NOT_ALLOWED
                  - 406 NOT_ACCEPTABLE
                  - 407 PROXY_AUTHENTICATION_REQUIRED
                  - 408 REQUEST_TIMEOUT
                  - 409 CONFLICT
                  - 410 GONE
                  - 411 LENGTH_REQUIRED
                  - 412 PRECONDITION_FAILED
                  - 413 PAYLOAD_TOO_LARGE
                  - 413 REQUEST_ENTITY_TOO_LARGE
                  - 414 URI_TOO_LONG
                  - 414 REQUEST_URI_TOO_LONG
                  - 415 UNSUPPORTED_MEDIA_TYPE
                  - 416 REQUESTED_RANGE_NOT_SATISFIABLE
                  - 417 EXPECTATION_FAILED
                  - 418 I_AM_A_TEAPOT
                  - 419 INSUFFICIENT_SPACE_ON_RESOURCE
                  - 420 METHOD_FAILURE
                  - 421 DESTINATION_LOCKED
                  - 422 UNPROCESSABLE_ENTITY
                  - 423 LOCKED
                  - 424 FAILED_DEPENDENCY
                  - 425 TOO_EARLY
                  - 426 UPGRADE_REQUIRED
                  - 428 PRECONDITION_REQUIRED
                  - 429 TOO_MANY_REQUESTS
                  - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
                  - 451 UNAVAILABLE_FOR_LEGAL_REASONS
                  - 500 INTERNAL_SERVER_ERROR
                  - 501 NOT_IMPLEMENTED
                  - 502 BAD_GATEWAY
                  - 503 SERVICE_UNAVAILABLE
                  - 504 GATEWAY_TIMEOUT
                  - 505 HTTP_VERSION_NOT_SUPPORTED
                  - 506 VARIANT_ALSO_NEGOTIATES
                  - 507 INSUFFICIENT_STORAGE
                  - 508 LOOP_DETECTED
                  - 509 BANDWIDTH_LIMIT_EXCEEDED
                  - 510 NOT_EXTENDED
                  - 511 NETWORK_AUTHENTICATION_REQUIRED
...
  /result/{test_id}:
    get:
      tags:
        - result-controller
      summary: Получить результаты теста
      description: Возвращает результаты прохождения теста для указанного test_id.
      operationId: getResult
      parameters:
        - name: test_id
          in: path
          description: Идентификатор теста
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успешное получение результатов теста
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultResponseDTO'
        '404':
          description: Результаты не найдены
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultResponseDTO'
...
components:
  schemas:
...
    CommentEditRequestDTO:
      type: object
      properties:
        newComment:
          type: string
          description: новый комментарий
...
    AnswerTestModel:
      type: object
      properties:
        value:
          type: integer
          description: Значение ответа (0-4)
          format: int32
          example: 3
        comment:
          type: string
          description: Комментарий к ответу
          example: Отличный вопрос, но не совсем понял формулировку
      description: Ответ на тестовый вопрос
    ResultRequestDTO:
      type: object
      properties:
        evaluatedId:
          type: integer
          description: ID оцениваемого пользователя
          format: int32
          example: 456
        evaluatorId:
          type: integer
          description: ID оценивающего пользователя
          format: int32
          example: 789
        startTimeStamp:
          type: integer
          description: Время начала теста (Unix timestamp)
          format: int64
          example: 1710567890
        endTimeStamp:
          type: integer
          description: Время окончания теста (Unix timestamp)
          format: int64
          example: 1710571490
        answers:
          type: array
          description: Ответы на тестовые вопросы
          items:
            $ref: '#/components/schemas/AnswerTestModel'
        skills:
          type: array
          description: Оценки навыков
          items:
            $ref: '#/components/schemas/SkillsTestModel'
      description: Отправка результатов после теста
    SkillsTestModel:
      type: object
      properties:
        skillId:
          type: integer
          description: ID навыка
          format: int32
          example: 12
        value:
          type: number
          description: Среднее значение оценки навыка
          format: double
          example: 3.8
      description: Оценка навыков в тесте
...
    ResultResponseDTO:
      type: object
      properties:
        title:
          type: string
          description: Название теста
          example: Оценка командной работы
        averageResult:
          type: number
          description: Средний результат
          format: double
          example: 4.2
        thisResult:
          type: number
          description: Личный результат
          format: double
          example: 4
        commandResult:
          type: number
          description: Результат команды
          format: double
          example: 4.3
        expertResult:
          type: number
          description: Результат экспертов
          format: double
          example: 4.1
        results:
          type: array
          description: Результаты по навыкам
          items:
            $ref: '#/components/schemas/SkillsResultModel'
        resultComment:
          type: string
          description: Финальный комментарий
          example: В целом хороший результат, но есть куда расти
      description: Результаты теста
    SkillsResultModel:
      type: object
      properties:
        skillName:
          type: string
          description: Название навыка
          example: Командная работа
        averageEvaluation:
          type: number
          description: Средняя оценка по навыку
          format: double
          example: 4.2
        thisEvaluation:
          type: number
          description: Личная оценка
          format: double
          example: 4
        commandEvaluation:
          type: number
          description: Оценка команды
          format: double
          example: 4.3
        expertEvaluation:
          type: number
          description: Оценка эксперта
          format: double
          example: 4.1
        comments:
          type: array
          description: Комментарии
          example:
            - Очень хорошо работает в команде
            - Немного медлит в принятии решений
          items:
            type: string
            description: Комментарии
            example: '["Очень хорошо работает в команде","Немного медлит в принятии решений"]'
      description: Результаты по навыкам
...